[BUS-1900] update version on master branch

+++ admin/install_shared.sh
@@ -10,15 +10,17 @@ ln -s ${INSTALL_BASE}/${PROJECT_NAME}/${VERSION}-${BUILD_NUMBER} ${INSTALL_BASE}
 
 #####################################################################################
 # create tag in git
-BUILD="${VERSION}-${BUILD_NUMBER}"
+if [ "${VERSION}" != "master" ]; then
+   BUILD="${VERSION}-${BUILD_NUMBER}"
 
-# remove tag from remote; retag and push tag to remote
-git push origin :refs/tag/${BUILD}
-git tag -f -a ${BUILD} -m "[${BUILD}]: auto-create tag for build" && git push -f origin ${BUILD}
-rc=$?
-if [ $rc != 0 ]; then
-   echo " Error ($rc): Could not create git tag for ${PROJECT_NAME}-${BUILD}}"
-   exit $rc;
+   # remove tag from remote; retag and push tag to remote
+   git push origin :refs/tag/${BUILD}
+   git tag -f -a ${BUILD} -m "[${BUILD}]: auto-create tag for build" && git push -f origin ${BUILD}
+   rc=$?
+   if [ $rc != 0 ]; then
+      echo " Error ($rc): Could not create git tag for ${PROJECT_NAME}-${BUILD}}"
+      exit $rc;
+   fi
 fi
 
 exit 0
